Переносит BLANK_POINT в файл констант
